edited some file and function names
