+++ src/js/routes/feedPage.js
@@ -1,21 +1,21 @@
-import * as listener from "../ui/listeners/index.js";
+import * as listenFor from "../ui/listeners/index.js";
 import * as templates from "../templates/posts/index.js";
 import * as HTTPMethod from "../api/requests/index.js";
 
-async function buildFeed(allPosts) {
+async function startFeed(allPosts) {
   const feedContainer = document.querySelector(".feed-container");
   feedContainer.innerHTML = "";
   templates.renderPostTemplates(allPosts, feedContainer);
-  await listener.openModal();
+  await listenFor.openPostAsModal();
 }
 
 export async function feedPage() {
   try {
     const posts = await HTTPMethod.getPostsFromAPI();
     if (posts) {
-      await buildFeed(posts);
-      listener.search(posts);
-      listener.postListener();
+      await startFeed(posts);
+      listenFor.search(posts);
+      listenFor.publishNewPost();
     } else if (!posts) {
       throw new Error("something went wrong when calling API");
     }
