+++ src/js/ui/events/index.js
@@ -1,7 +1,7 @@
 //this is a barrel-file for all the js files in the posts-folder (for export purpose) - from Olis YT tutorial
 
 export * from "./auth.js";
-export * from "./newPostContent.js";
+export * from "./newPost.js";
 export * from "./updatePostContent.js";
 export * from "./search.js";
 export * from "./filter.js";
