+++ src/js/ui/events/newPost.js
@@ -1,6 +1,6 @@
 import { sendPostToAPI } from "../../api/requests/index.js";
 
-export async function newPostContent(event) {
+export async function newPost(event) {
   event.preventDefault();
 
   const form = event.target;
