+++ src/js/ui/listeners/index.js
@@ -1,6 +1,8 @@
 //this is a barrel-file for all the js files in the posts-folder (for export purpose) - from Olis YT tutorial
 
 export * from "./onAuth.js";
-export * from "./onPost.js";
+export * from "./publishPost.js";
 export * from "./onSearch.js"
+export * from "./openModal.js"
+
