+++ src/js/ui/listeners/onAuth.js
@@ -1,6 +1,6 @@
 import { loginAuth, registerAuth } from "../events/index.js";
 
-export function authListener() {
+export function authentication() {
   document.forms.login.addEventListener("submit", loginAuth);
   document.forms.register.addEventListener("submit", registerAuth);
 }
