+++ src/js/ui/listeners/publishPost.js
@@ -1,0 +1,5 @@
+import { newPost } from "../events/index.js";
+
+export function publishNewPost() {
+  document.forms.newPost.addEventListener("submit", newPost);
+}
